Moved Meter stories to separate files
