+++ src/js/components/Meter/stories/Bar.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box, Meter } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+const BarMeter = () => (
+  <Grommet theme={grommet}>
+    <Box align="center" pad="large">
+      <Meter type="bar" background="light-2" values={[{ value: 30 }]} />
+    </Box>
+  </Grommet>
+);
+
+storiesOf('Meter', module).add('Bar', () => <BarMeter />);
